refactor(relationship): 将用户 ID 类型从 Integer 改为 Long
- 修改了 FriendApplyDo 和 UserFriendDo 类中的用户 ID 字段类型
- 更新了相关 mapper 和 service 接口中的方法参数和返回值类型
-移除了不必要的依赖项，精简了 pom.xml 文件

+++ spring-server/api/api-model/src/main/java/com/czy/api/domain/Do/relationship/UserFriendDo.java
@@ -9,9 +9,9 @@ import lombok.Data;
  */
 @Data
 public class UserFriendDo implements BaseBean {
-    private Integer id;
-    private Integer userId;
-    private Integer friendId;
+    private Long id;
+    private Long userId;
+    private Long friendId;
     private Long addTime = System.currentTimeMillis();
     private Long lastChatTime;
 }
